Implement full NVML GPU monitoring, graph rendering, and inference engine status tracking
- Full NVML collector using nvml-wrapper crate with GPU utilization, memory, temperature, power, clocks, fan speed, PCIe throughput, and processes
- Graph widgets for GPU history (utilization, memory, temperature, power)
- Graph widgets for CPU history and inference metrics history
- Sparkline rendering in overview view
- Inference engine status tracking (Connected/Unavailable/Connecting)
- Ollama and vLLM engines return status along with models
- Inference collector updates engine status on each collect cycle
- Fixed ratatui 0.29 API (Borders::all(), Terminal API)
- Fixed sysinfo API (load_average, static methods)
- Fixed model serialization (removed Serialize from Instant-containing structs)

+++ src/models/gpu.rs
@@ -24,8 +24,8 @@ pub struct GpuMetrics {
     pub sm_clock: Option<u32>,
     pub mem_clock: Option<u32>,
     pub fan_speed: Option<u32>,
-    pub pcie_tx: Option<u64>,
-    pub pcie_rx: Option<u64>,
+    pub pcie_tx: Option<u32>,
+    pub pcie_rx: Option<u32>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
